requirements.txt
beautifulsoup4
pandas
selenium
schedule
webdriver-manager
mysql-connector-python